+++ backend/src/Constant/Package/PackageConstant.php
@@ -7,5 +7,4 @@ final class PackageConstant
     const PACKAGE_ACTIVE = "active";
 
     const PACKAGE_INACTIVE = "inactive";
-
 }
