+++ backend/src/Entity/StoreOwnerProfileEntity.php
@@ -1,7 +1,7 @@
 <?php
 
 namespace App\Entity;
-use App\Repository\StoreOwner\StoreOwnerProfileEntityRepository;
+use App\Repository\StoreOwnerProfileEntityRepository;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
