+++ backend/src/Entity/UserEntity.php
@@ -2,7 +2,7 @@
 
 namespace App\Entity;
 
-use App\Repository\User\UserEntityRepository;
+use App\Repository\UserEntityRepository;
 use Doctrine\ORM\Mapping as ORM;
 use Symfony\Component\Security\Core\User\PasswordAuthenticatedUserInterface;
 use Symfony\Component\Security\Core\User\UserInterface;
